requirements.txt
requests
supabase
python-dotenv
pandas
streamlit
plotly
psycopg2-binary
apscheduler
